extended work (desc)
-light touchup work done on boat char
just to make some stuff accessible in BP where needed as well as making the rotation behaviour a bit more consistent. light tweaks to other things on the uasset too.
-sunseeker reimport
scaling was off, needed to retweak it a few times. turns out a sunseeker is quite small based on ref footage!
-HUD class made
just needs to be blueprint to handle some basic stuff; function added to spawn a widget and take control from the player when they win
-win widget added
player can either restart the game or close the game after winning
-trigger made
this is all handled in BP and checks if the player is in the right rotation and is slow enough to be considered docking. this is a generous amount, but with how the player slows down, it's easy to shoot out.
-boat spline readded with a CPP base class
-boat obstacle BP work
the boat obstacles now run according to timelines that can be modified, meaning they will travel along the splines at whatever speed you like, exposed to the editor so you can change things around as desired. all will travel around their spline courses over the course of a minute.
-boat obstacle CPP work
added a function exposed to BP that handles the movement of the actual obstacles along the spline. this could've also been done in blueprint but I wanted to show that it could be done in CPP and exposed to blueprint via a custom node
-particle added
this is spawned and removed when you're in the right rotation / speed on the dock trigger.
-additional misc changes to level, materials